requirements.txt
asgiref==3.6.0
Django==4.1.4
psycopg2==2.9.5
python-dotenv==0.21.0
sqlparse==0.4.3
